feat(chakra-components): atomizing components, part 2
