+++ packages/chakra-components/src/components/Election/index.tsx
@@ -10,3 +10,4 @@ export * from './SpreadsheetAccess'
 export * from './StatusBadge'
 export * from './Title'
 export * from './VoteButton'
+export * from './VoteWeight'
