+++ packages/chakra-components/src/components/Organization/Header.tsx
@@ -12,3 +12,4 @@ export const OrganizationHeader = (props: IPFSImageProps) => {
 
   return <Image src={header} sx={styles} {...props} />
 }
+OrganizationHeader.displayName = 'OrganizationHeader'
