+++ packages/chakra-components/src/components/Pagination/Pagination.tsx
@@ -30,12 +30,14 @@ const PageButton = ({ page, currentPage, ...rest }: PaginatorButtonProps) => (
     {page + 1}
   </PaginatorButton>
 )
+PageButton.displayName = 'PageButton'
 
 const RoutedPageButton = ({ page, currentPage, to, ...rest }: PaginatorButtonProps & { to: string }) => (
   <PaginatorButton as={RouterLink} to={to} isActive={currentPage === page} {...rest}>
     {page + 1}
   </PaginatorButton>
 )
+RoutedPageButton.displayName = 'RoutedPageButton'
 
 type CreatePageButtonType = (i: number) => ReactElement
 type GotoPageType = (page: number) => void
@@ -171,6 +173,7 @@ export const Pagination = ({ maxButtons = 10, buttonProps, inputProps, paginatio
     />
   )
 }
+Pagination.displayName = 'Pagination'
 
 export const RoutedPagination = ({ maxButtons = 10, buttonProps, pagination, ...rest }: PaginationProps) => {
   const { getPathForPage, setPage, page } = useRoutedPagination()
@@ -191,3 +194,4 @@ export const RoutedPagination = ({ maxButtons = 10, buttonProps, pagination, ...
     />
   )
 }
+RoutedPagination.displayName = 'RoutedPagination'
