+++ packages/chakra-components/src/components/Pagination/PaginatorButton.tsx
@@ -4,3 +4,4 @@ export const PageButton = forwardRef<ButtonProps, 'div'>((props, ref) => {
   const styles = useStyleConfig('PageButton', props)
   return <Button sx={styles} {...props} />
 })
+PageButton.displayName = 'PageButton'
